requirements.txt
# LiveKit Agents Framework
livekit-agents
livekit

# AI Model Plugins
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-elevenlabs
livekit-plugins-silero
livekit-plugins-noise-cancellation

# Utilities
python-dotenv
aiohttp

# Telemetry & Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http

